Build: Transpile dependencies that otherwise break the build (#89)
At some point the `jsonpointer` library started shipping code with
ES2015+ syntax and `const` was breaking `UglifyJS`. This has resulted
in an inability to build the project as described in the documention.

In this patch we're adding that module to our list of sources for
`babel` to transpile, removing that syntax which `UglifyJS` doesn't
understand.