Introduce package lock to prevent future breakage on dependency changes (#92)
In #89 we had to make a fix to restore the ability to build this project
because a dependency started shipping untranspiled code. This could have
been prevented if we had been using a `package-lock.json` file and retained
the old dependency version.

In this patch we're introducing a package lock based on the current project
dependencies to prevent such a future problem from appearing, inasmuch as
we are able to control the parocess.